% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holding.R
\name{get_holding}
\alias{get_holding}
\title{read_holding}
\usage{
get_holding(
  etf = c("innovation", "genomic", "fintech", "space", "israel", "internet", "robot",
    "printing")
)
}
\description{
read_holding
}
