\name{getVotes}
\alias{getVotes}
\title{Get Votes}
\usage{
getVotes(address)
}
\description{
Returns the delegate information for the delegate the account has voted for and returns a list
}
\examples{
getVotes("ARAq9nhjCxwpWnGKDgxveAJSijNG8Y6dFQ")

getVotes("ARAq9nhjCxwpWnGKDgxveAJSijNG8Y6dFQ")$delegates[[1]]$username

}
